+++ CHANGELOG.md
@@ -10,6 +10,7 @@ in 2.x, 1.x and 0.x versions.
   * Added part unit price field for accessories
   * Added bars and sheets scrap's cross display / hide in cutting diagram options
   * Added cutting diagram 2D option to select origin corner - top left or bottom left
+  * Added comments display on news
   
 * 1.9.10 (2020-12-17)